PO-9347 Integrate Find Bugs into PA build process as another static code analysis tool.
git-svn-id: https://ncisvn.nci.nih.gov/svn/coppa/trunk/code@19131 553898ca-32d5-487a-af8b-6cf4ca579ab4

+++ pa/pa-ejb/src/java/gov/nih/nci/pa/service/ParticipatingSiteBeanLocal.java
@@ -847,8 +847,10 @@ public class ParticipatingSiteBeanLocal extends AbstractParticipatingSitesBean /
         if (src == null || dest == null) {
             throw new PAException("Site not found when merging participating sites.");
         }
-        if (src.getStudyProtocol().getId() != dest.getStudyProtocol().getId()) {
-            throw new PAException("Trying to merge participating sites from different trials.");
+        if (src.getStudyProtocol().getId().longValue() != dest
+                .getStudyProtocol().getId().longValue()) {
+            throw new PAException(
+                    "Trying to merge participating sites from different trials.");
         }
         mergeAccrualCounts(src, dest);
         mergeAccrualSubjects(src, dest);
